grpc-gke-demo/requirements.txt
grpcio==1.64.0
grpcio-tools==1.64.0
google-cloud-container
google-cloud-logging
